moves encoding to own pkg
